Makefile: Add the install-tools target
ith a fresh checked out k3d tree without required tools installed, Any make command gives the following error message:

    Makefile:71: *** recipe commences before first target.  Stop.

Basically, the section around line 71 are not properly formed makefile rules. This error is hidden when required tools
are installed. The error can be re-exposed with the "make extra-clean" command.

The install-tools target is intended to be invoked when installing the development tree for the fist time.