Makefile: Avoid setting k3s version at build time
Commit:
	commit e82e95a23d37d50e0c44ae68aa92bbe5f2f4b877
	Author: Chris Carty <chriscartydev@gmail.com>
	Date:   Fri May 3 22:04:39 2019 -0400

	update default K3sVersion to use latest tag

Made the change to use tag "latest" for k3s images by default.

However, without this commit, the default tag will be set to a tag that
was set at the k3d build time.  It is a bit odd to set run time value at build
time in general.  This commit removes the build time setting.