Add --no-remove/--no-rm option
When a k3s cluster is delete, the container images are not automatically
removed. Overtime, this can cause the file system used by the docker daemon
to fill up.  This patch sets up the auto remove flag when creating k3s
containers, so they are deleted automatically when done.

In case the container images are useful (mostly for debugging purpose),
this patch adds the --no-remove flag.