feat: Allow creating the registry on user-defined networks
This might be useful when a user wants to use a different network to run
the registry, especially for Podman users.

This patch avoids adding such a flag to clusterCreate to avoid polluting
the arguments for that command. A better long-term solution would be to
create a new network for the registry.