plugin: List k3d plugins
Add a new command `k3d plugin list` that lists plugins inside a folder,
by default it is `$HOME/.k3d/plugins`.

Every folder inside the plugins directory is considered a plugin and it
must contain a manifest named `k3d-plugin.yaml`. This file describes a
plugin, so it must have a name, a version, a description and an help
message.

Here's an example of a plugin manifest:
```yaml
    ---
    name: env
    version: "0.0.1"
    shortHelpMessage: "Print k3d env variables"
    helpMessage: "Print k3d env variables"
    command: <command to list env variables>
```