fix: use LB for node registration otherwise fallback to server node
This commit is used to fix the error when server-0 is deleted from
the cluster. Any subsequent nodes that are added results in error
since server-0 is considered registration node by default. Also,
added a tls-san for server-lb to avoid k8s fatal error for mistmatched
hostname when registering a server node using server-lb the first time.