Adhere to the XDG specification for configuration
So far K3d has been creating the '.k3d' directory on the user's 'HOME'
directory, which isn't the ideal place for configuration and violates
the XDG base directory specification.

This commit aligns the configuration persistency with the XDG spec,
making the 'GetConfigDirOrCreate' function obtain the config dir
from the 'XDG_CONFIG_HOME' environment variable, falling back to the
'~/.config/k3d' default path in case the former is not set.

Signed-off-by: Carlos Torres <torr.ca@mailgw.com>