chore: Update the hardcoded k3s version from 1.21 to 1.33
This is overridden at build time in the Makefile but some distros might not use that and fall back to the default version.
This is the case for example for Nix where the user _can_ override the version but if that doesn't happen they default to a very old one (from 2021 or so).

https://github.com/NixOS/nixpkgs/blob/0d36547deb33d32f05add583384ae5858dadb464/pkgs/by-name/k3/k3d/package.nix#L6

This just updates the default version to the current stable k3s version.